Updated Gradle to 5.4 and migrated to AndroidX.
I also changed the target API level to API 23 (Marshmallow) due to it
being more proper for a target API level (as if statements won't have to
be made to properly integrate permissions as permissions start being
implemented in the app (such as read/write storage)), and I also decided
to .gitignore the .idea folder as it is not necessary and it has the
possibility of messing up over time due to different configurations
by users who may contribute to the project.

Made some other changes as well, such as improving ArticleAdapter's
class initializer by stopping it from being all one line.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>